Create hooks dir to separate the useEvents and useSocket that were in the App
